Main object-orientated skeleton of code
This is the initial skeleton of the solution, with some of the needed code. The png_group object will contain multiple png objects, and the svg_group object will contain multiple svg objects. The png and svg objects only contain the raw image data. Many of the conversion methods (to bitmap, to greyscale, etc) are contained in the png_group and svg_group classes. The final_create class contains the from_session(session) method that the user will call and pass a session object to, and that class will also do the final combination of svg and png images.