Add files via upload
Integrated the code with Tangwa's. Added method to remove the transparency of the images that need to be vectorized (potrace seemed to vectorize the entire image if the background was transparent). Still struggling with the final output image, in that the items are misplaced. The contours don't automatically align with the background image, and many of the items on the side of the image (colourbar, title, numbers, border, etc) end up being drawn over one another.